+++ Scribe/Scribe-Api/src/utopia/scribe/api/database/factory/logging/StackTraceElementRecordFactory.scala
@@ -2,8 +2,8 @@ package utopia.scribe.api.database.factory.logging
 
 import utopia.flow.generic.model.immutable.Model
 import utopia.scribe.api.database.ScribeTables
-import utopia.scribe.core.model.partial.logging.StackTraceElementData
-import utopia.scribe.core.model.stored.logging.StackTraceElement
+import utopia.scribe.core.model.partial.logging.StackTraceElementRecordData
+import utopia.scribe.core.model.stored.logging.StackTraceElementRecord
 import utopia.vault.nosql.factory.row.model.FromValidatedRowModelFactory
 
 /**
@@ -11,16 +11,16 @@ import utopia.vault.nosql.factory.row.model.FromValidatedRowModelFactory
   * @author Mikko Hilpinen
   * @since 22.05.2023, v0.1
   */
-object StackTraceElementFactory extends FromValidatedRowModelFactory[StackTraceElement]
+object StackTraceElementRecordFactory extends FromValidatedRowModelFactory[StackTraceElementRecord]
 {
 	// IMPLEMENTED	--------------------
 	
 	override def defaultOrdering = None
 	
-	override def table = ScribeTables.stackTraceElement
+	override def table = ScribeTables.stackTraceElementRecord
 	
 	override protected def fromValidatedModel(valid: Model) = 
-		StackTraceElement(valid("id").getInt, StackTraceElementData(valid("className").getString, 
+		StackTraceElementRecord(valid("id").getInt, StackTraceElementRecordData(valid("className").getString,
 			valid("methodName").getString, valid("lineNumber").getInt, valid("causeId").int))
 }