+++ Scribe/Scribe-Core/data/sql/scribe-db-structure-v0.1.sql
@@ -3,6 +3,8 @@
 -- Version: v0.1
 -- Last generated: 2023-05-22
 --
+-- TODO: Must be rebuilt
+--
 
 --	Logging	----------